+++ StabilityMatrix.Avalonia/ViewModels/PackageManagerViewModel.cs
@@ -6,8 +6,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using AsyncAwaitBestPractices;
 using Avalonia.Controls;
-using Avalonia.Controls.Notifications;
-using Avalonia.Controls.Primitives;
 using Avalonia.Threading;
 using CommunityToolkit.Mvvm.Input;
 using DynamicData;
@@ -15,18 +13,15 @@ using DynamicData.Binding;
 using FluentAvalonia.UI.Controls;
 using Microsoft.Extensions.Logging;
 using StabilityMatrix.Avalonia.Animations;
-using StabilityMatrix.Avalonia.Controls;
+using StabilityMatrix.Avalonia.Languages;
 using StabilityMatrix.Avalonia.Services;
 using StabilityMatrix.Avalonia.ViewModels.Base;
-using StabilityMatrix.Avalonia.ViewModels.Dialogs;
 using StabilityMatrix.Avalonia.ViewModels.PackageManager;
 using StabilityMatrix.Avalonia.Views;
-using StabilityMatrix.Avalonia.Views.Dialogs;
 using StabilityMatrix.Core.Attributes;
 using StabilityMatrix.Core.Helper;
 using StabilityMatrix.Core.Models;
 using StabilityMatrix.Core.Models.FileInterfaces;
-using StabilityMatrix.Core.Models.PackageModification;
 using StabilityMatrix.Core.Models.Packages;
 using StabilityMatrix.Core.Services;
 using Symbol = FluentIcons.Common.Symbol;
@@ -48,7 +43,7 @@ public partial class PackageManagerViewModel : PageViewModelBase
     private readonly INavigationService<NewPackageManagerViewModel> packageNavigationService;
     private readonly ILogger<PackageManagerViewModel> logger;
 
-    public override string Title => "Packages";
+    public override string Title => Resources.Label_Packages;
     public override IconSource IconSource => new SymbolIconSource { Symbol = Symbol.Box, IsFilled = true };
 
     /// <summary>
@@ -84,6 +79,7 @@ public partial class PackageManagerViewModel : PageViewModelBase
         this.logger = logger;
 
         EventManager.Instance.InstalledPackagesChanged += OnInstalledPackagesChanged;
+        EventManager.Instance.OneClickInstallFinished += OnOneClickInstallFinished;
 
         var installed = installedPackages.Connect();
         var unknown = unknownInstalledPackages.Connect();
@@ -107,6 +103,11 @@ public partial class PackageManagerViewModel : PageViewModelBase
         timer.Tick += async (_, _) => await CheckPackagesForUpdates();
     }
 
+    private void OnOneClickInstallFinished(object? sender, bool e)
+    {
+        OnLoadedAsync().SafeFireAndForget();
+    }
+
     public void SetPackages(IEnumerable<InstalledPackage> packages)
     {
         installedPackages.Edit(s => s.Load(packages));
@@ -119,7 +120,7 @@ public partial class PackageManagerViewModel : PageViewModelBase
 
     public override async Task OnLoadedAsync()
     {
-        if (Design.IsDesignMode)
+        if (Design.IsDesignMode || !settingsManager.IsLibraryDirSet)
             return;
 
         installedPackages.EditDiff(settingsManager.Settings.InstalledPackages, InstalledPackage.Comparer);