+++ StabilityMatrix.Avalonia/Views/CivitAiBrowserPage.axaml.cs
@@ -1,8 +1,11 @@
-﻿using System.Diagnostics;
+﻿using System;
+using System.Diagnostics;
+using AsyncAwaitBestPractices;
 using Avalonia.Controls;
 using Avalonia.Input;
 using Avalonia.Markup.Xaml;
 using StabilityMatrix.Avalonia.Controls;
+using StabilityMatrix.Avalonia.Models;
 using StabilityMatrix.Core.Attributes;
 using CivitAiBrowserViewModel = StabilityMatrix.Avalonia.ViewModels.CheckpointBrowser.CivitAiBrowserViewModel;
 
@@ -26,8 +29,15 @@ public partial class CivitAiBrowserPage : UserControlBase
         if (sender is not ScrollViewer scrollViewer)
             return;
 
-        var isAtEnd = scrollViewer.Offset == scrollViewer.ScrollBarMaximum;
-        Debug.WriteLine($"IsAtEnd: {isAtEnd}");
+        if (scrollViewer.Offset.Y == 0)
+            return;
+
+        var isAtEnd = Math.Abs(scrollViewer.Offset.Y - scrollViewer.ScrollBarMaximum.Y) < 1f;
+
+        if (isAtEnd && DataContext is IInfinitelyScroll scroll)
+        {
+            scroll.LoadNextPageAsync().SafeFireAndForget();
+        }
     }
 
     private void InputElement_OnKeyDown(object? sender, KeyEventArgs e)