+++ StabilityMatrix.Core/Models/Api/CivitBaseModelType.cs
@@ -8,6 +8,9 @@ public enum CivitBaseModelType
 {
     All,
 
+    [StringValue("Pony")]
+    Pony,
+
     [StringValue("SD 1.5")]
     Sd15,
 
@@ -26,7 +29,20 @@ public enum CivitBaseModelType
     [StringValue("SDXL 1.0 LCM")]
     Sdxl10Lcm,
 
+    [StringValue("SDXL Distilled")]
+    SdxlDistilled,
+
+    [StringValue("SDXL Lightning")]
+    SdxlLightning,
+
     [StringValue("SDXL Turbo")]
     SdxlTurbo,
+
+    [StringValue("SVD")]
+    SVD,
+
+    [StringValue("Stable Cascade")]
+    StableCascade,
+
     Other,
 }