+++ StabilityMatrix.Core/Models/Packages/Extensions/IPackageExtensionManager.cs
@@ -95,6 +95,12 @@ public interface IPackageExtensionManager
     Task<IEnumerable<InstalledPackageExtension>> GetInstalledExtensionsLiteAsync(
         InstalledPackage installedPackage,
         CancellationToken cancellationToken = default
+
+    /// <summary>
+    /// Get updated info (version) for an installed extension.
+    /// </summary>
+    Task<InstalledPackageExtension> GetInstalledExtensionInfoAsync(
+        InstalledPackageExtension installedExtension
     );
 
     /// <summary>