+++ StabilityMatrix.Core/Models/Settings/Settings.cs
@@ -134,6 +134,7 @@ public class Settings
     public Size OutputsImageSize { get; set; } = new(300, 300);
     public HolidayMode HolidayModeSetting { get; set; } = HolidayMode.Automatic;
     public bool IsWorkflowInfiniteScrollEnabled { get; set; } = true;
+    public bool IsOutputsTreeViewEnabled { get; set; } = true;
 
     [JsonIgnore]
     public bool IsHolidayModeActive =>